requirements.txt
# Core ML Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# NLP
nltk==3.8.1

# MLOps Tools
dvc==3.27.0
mlflow==2.8.1

# Utilities
pyyaml==6.0.1
joblib==1.3.2

# Optional but recommended
matplotlib==3.7.2
seaborn==0.12.2